impl: Add backend code for the login page.
added user.js to controllers directory to manager user login. Added auth.js to middleware direcotry to ensure a non-logged in user cant use the website fully. Added user.js in models to prepare the Schema to send to the database. Added tasks.js to routes to help map out the changes to our application. Also added users.js for the same reason. Added Auth.js to components to handle the submit on the front end. Added auth.js to Actions to handle user actions. Issue #11