impl: Add Code For The Dashboard Backend
Added tasks.js in the controllers folder and added the getTasks function to it which gets the tasks from the database. Added taskMessage.js in the models folder which contains the schema for the tasks database. Changed router.get in the tasks.js file in the routes folder to refer to the getTasks function in the tasks.js file in the controllers folder, in order to be able to get the tasks from the database. Modified index.js to add the code to use the routes from the routes/tasks.js file which allows us to pull the data from the databse, and also allows the frontend to access the data by going to the /tasks page that's been specified. GitHub Issue #25.