impl: Add Code For The Task Manager
Added styles.js under the TaskDetail folder, and added styles.js and tasks.js in the TaskDetail/Task folder. Also modified TaskDetail.js to contain the code for the display of the tasks on the TaskDetail page, the styles of which are contained in /TaskDetail/styles.js. This code also contains the newly added Active and Todo buttons, which toggle the Active and Todo status of the tasks. Also added card-specific formatting code to styles.js and tasks.js under the TaskDetail/Task folder.GitHub Issue #28.