impl: Add Code For The Task Manager Buttons Functionality
Modified TaskDetail.js to contain the code for the display of the tasks on the TaskDetail page, the styles of which are contained in /TaskDetail/styles.js in such a way that each task goes under its appropriate column. Also modified TaskDetail/Task/tasks.js to appropriately format the card according to its status tag (active, todo, completed). GitHub Issue #28.