impl: Add burger menu functionality to dashboard. fix: Remove category from dropdown if task is listed as todo or completed. fix: Category dropdown will only show tasks that are not todo or completed
Added burger menu components to Dashboard.js. Created logic to ignore tasks that are todo or complete when creating the category dropdown. Added Burger.js to control layout of the menu and actions to the burger. Added sort/filter form to Menu.js to later be made functional. Added Menu.style.js and Burger.style.js for styling the components. Added index.js to Burger and Menu folders to allow exporting of the components. Installed 'styled-components' to allow for themes to be added to the burger and menu. In tasks.js the page will now be reloaded if the user clicks todo or complete on a task. This allows dashboard to reevaluate if a task category has a status of todo or completed, and then does not add that category to the list Issue: #31