Imple: Connected the google auth to the Score
When a use completes a task it checks to see if a user with a matching email as the google email. if there is then update that score counter if there is not then create a new user and set the score to 1
Since there is no way to fill in the user password this way i fill in the value as temp. Then when the user goes to sign without the google signin if they fill in a password that matches all the required restrictions set in the front end then we update the password with this new one.

Github issue #37